requirements.txt
requests==2.32.5
aiohttp==3.12.15
pymysql==1.1.2
python-dotenv==1.1.1
pydantic==2.10.4
pydantic-settings==2.7.0
schedule==1.2.0